# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is an application web of interview assessment tests for interviewing out of the box .",
        "title": "test",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/my-tests/questions/delete": {
            "delete": {
                "description": "delete a question from test by json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question_test"
                ],
                "summary": "delete a question from test",
                "parameters": [
                    {
                        "description": "test_question",
                        "name": "test_question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestQuestionDelete"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/my-tests/{id}/questions": {
            "post": {
                "description": "add a question to test by json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question_test"
                ],
                "summary": "add a question to test",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add question to test",
                        "name": "test_question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestQuestion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestQuestion"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.TestQuestion": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                }
            }
        },
        "models.TestQuestionDelete": {
            "type": "object",
            "required": [
                "question_id",
                "test_id"
            ],
            "properties": {
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}